HW06: Refactor tests
