+++ hw06_testing/hw03/internal/reader/reader_test.go
@@ -1,0 +1,13 @@
+package reader
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestConsoleReader(t *testing.T) {
+	result := ConsoleReader()
+
+	assert.Equal(t, 8, result)
+}
