+++ hw06_testing/hw05/pkg/shape/types/shapes/shape_test.go
@@ -43,7 +43,7 @@ func TestShape_CalcSquare(t *testing.T) {
 	}
 }
 
-func TestShape_CalcCircle(t *testing.T) {
+func TestShape_String(t *testing.T) {
 	tests := []struct {
 		name  string
 		shape fmt.Stringer
